+++ hil-test/tests/qspi.rs
@@ -211,9 +211,7 @@ mod tests {
 
         let spi = Spi::new(
             peripherals.SPI2,
-            Config::default()
-                .with_frequency(100.kHz())
-                .with_mode(Mode::_0),
+            Config::default().with_clock(100.kHz()).with_mode(Mode::_0),
         )
         .unwrap();